% tinyrox says don't edit this manually, but it can't stop you!
\name{$.glinty_input}
\alias{$.glinty_input}
\title{Access an input value}
\usage{
\method{$}{glinty_input}(x, name)
}
\arguments{
\item{x}{a glinty_input proxy}

\item{name}{character input ID}
}
\value{
a reactive_val function
}
\description{
Access an input value
}
