% tinyrox says don't edit this manually, but it can't stop you!
\name{a}
\alias{a}
\title{Create a hyperlink}
\usage{
a(text, href)
}
\arguments{
\item{text}{character link text}

\item{href}{character link target URL}
}
\value{
A UI element
}
\description{
Create a hyperlink
}
\examples{
a("cornball.ai", "https://cornball.ai")
}
