% tinyrox says don't edit this manually, but it can't stop you!
\name{add_pending_flush}
\alias{add_pending_flush}
\title{Add an observer to the pending flush queue}
\usage{
add_pending_flush(obs)
}
\arguments{
\item{obs}{an observer environment}
}
\description{
Add an observer to the pending flush queue
}
\keyword{internal}
