% tinyrox says don't edit this manually, but it can't stop you!
\name{b64url_decode}
\alias{b64url_decode}
\title{Decode one base64url segment}
\usage{
b64url_decode(x)
}
\arguments{
\item{x}{character segment}
}
\value{
raw vector, or NULL on malformed input
}
\description{
Decode one base64url segment
}
\keyword{internal}
