% tinyrox says don't edit this manually, but it can't stop you!
\name{body_scripts_html}
\alias{body_scripts_html}
\title{Render app scripts for the end of the body}
\usage{
body_scripts_html(head)
}
\arguments{
\item{head}{a page head spec, or NULL}
}
\value{
character HTML (empty string when there are no scripts)
}
\description{
Render app scripts for the end of the body
}
\keyword{internal}
