% tinyrox says don't edit this manually, but it can't stop you!
\name{cancel_timer}
\alias{cancel_timer}
\title{Cancel a scheduled timer}
\usage{
cancel_timer(id)
}
\arguments{
\item{id}{integer timer id from schedule_timer()}
}
\value{
invisible(NULL)
}
\description{
Cancel a scheduled timer
}
\keyword{internal}
