% tinyrox says don't edit this manually, but it can't stop you!
\name{check_conditions}
\alias{check_conditions}
\title{Validate and unclass a set of conditions}
\usage{
check_conditions(args, fn)
}
\arguments{
\item{args}{list of candidate conditions}

\item{fn}{character calling function name, for the error message}
}
\value{
an unnamed list of plain condition lists
}
\description{
Validate and unclass a set of conditions
}
\keyword{internal}
