% tinyrox says don't edit this manually, but it can't stop you!
\name{check_job}
\alias{check_job}
\title{Refuse anything that is not a job handle}
\usage{
check_job(job)
}
\arguments{
\item{job}{the supposed handle}
}
\value{
invisible(NULL)
}
\description{
Refuse anything that is not a job handle
}
\keyword{internal}
