% tinyrox says don't edit this manually, but it can't stop you!
\name{check_named_subset}
\alias{check_named_subset}
\title{Require names to come from a closed set}
\usage{
check_named_subset(x, allowed, what)
}
\arguments{
\item{x}{a named list}

\item{allowed}{character vector of valid names}

\item{what}{character argument name for the error}
}
\value{
invisible(NULL)
}
\description{
Require names to come from a closed set
}
\keyword{internal}
