% tinyrox says don't edit this manually, but it can't stop you!
\name{check_theme_number}
\alias{check_theme_number}
\title{Validate one numeric theme token}
\usage{
check_theme_number(x, what, min = 0, max = 64)
}
\arguments{
\item{x}{candidate value}

\item{what}{character token name for the error}

\item{min,max}{numeric bounds}
}
\value{
the value as a plain number
}
\description{
Validate one numeric theme token
}
\keyword{internal}
