% tinyrox says don't edit this manually, but it can't stop you!
\name{check_video_mime}
\alias{check_video_mime}
\title{Refuse a media type that is not video}
\usage{
check_video_mime(mime)
}
\arguments{
\item{mime}{character lowercase media type}
}
\value{
mime, invisibly refused by stop() when it is not video
}
\description{
Refuse a media type that is not video
}
\keyword{internal}
