% tinyrox says don't edit this manually, but it can't stop you!
\name{children_to_html}
\alias{children_to_html}
\title{Lower a list of children}
\usage{
children_to_html(children)
}
\arguments{
\item{children}{list of components}
}
\value{
character HTML
}
\description{
Lower a list of children
}
\keyword{internal}
