% tinyrox says don't edit this manually, but it can't stop you!
\name{clamp_progress}
\alias{clamp_progress}
\title{Hold a progress fraction inside [0, 1]}
\usage{
clamp_progress(value)
}
\arguments{
\item{value}{numeric}
}
\value{
numeric in [0, 1]
}
\description{
Hold a progress fraction inside [0, 1]
}
\keyword{internal}
