% tinyrox says don't edit this manually, but it can't stop you!
\name{collect_tree_ids}
\alias{collect_tree_ids}
\title{Collect element ids from an unclassed tag tree}
\usage{
collect_tree_ids(x)
}
\arguments{
\item{x}{an unclassed tag tree node}
}
\value{
character vector of ids
}
\description{
Collect element ids from an unclassed tag tree
}
\keyword{internal}
