% tinyrox says don't edit this manually, but it can't stop you!
\name{component}
\alias{component}
\title{Construct a component}
\usage{
component(type, ...)
}
\arguments{
\item{type}{character component name, present in COMPONENT_SCHEMA}

\item{...}{fields for this component}
}
\value{
a glinty_component
}
\description{
Construct a component
}
\keyword{internal}
