% tinyrox says don't edit this manually, but it can't stop you!
\name{cond_not}
\alias{cond_not}
\title{Negate a condition}
\usage{
cond_not(condition)
}
\arguments{
\item{condition}{a condition}
}
\value{
A glinty_condition
}
\description{
Negate a condition
}
\examples{
cond_not(input_is("backend", "openai"))
}
