% tinyrox says don't edit this manually, but it can't stop you!
\name{condition_json}
\alias{condition_json}
\title{Serialize a condition to JSON for the client}
\usage{
condition_json(condition)
}
\arguments{
\item{condition}{a glinty_condition}
}
\value{
character JSON
}
\description{
Serialize a condition to JSON for the client
}
\keyword{internal}
