% tinyrox says don't edit this manually, but it can't stop you!
\name{condition_matches}
\alias{condition_matches}
\title{Test one input value against a condition's candidates}
\usage{
condition_matches(actual, wanted)
}
\arguments{
\item{actual}{the input's current value}

\item{wanted}{list or vector of candidate values}
}
\value{
logical
}
\description{
Test one input value against a condition's candidates
}
\keyword{internal}
