% tinyrox says don't edit this manually, but it can't stop you!
\name{df_to_html}
\alias{df_to_html}
\title{Serialize a data.frame to an escaped HTML table}
\usage{
df_to_html(df)
}
\arguments{
\item{df}{a data.frame}
}
\value{
character HTML
}
\description{
Serialize a data.frame to an escaped HTML table
}
\keyword{internal}
