% tinyrox says don't edit this manually, but it can't stop you!
\name{divider}
\alias{divider}
\title{Create a horizontal rule}
\usage{
divider(label = NULL)
}
\arguments{
\item{label}{character text shown in the rule, e.g. "or"}
}
\value{
A UI component
}
\description{
Create a horizontal rule
}
\examples{
divider()
divider(label = "or")
}
