% tinyrox says don't edit this manually, but it can't stop you!
\name{drain_all_sessions}
\alias{drain_all_sessions}
\title{Drain every live session's outgoing queue}
\usage{
drain_all_sessions()
}
\value{
invisible(NULL)
}
\description{
Drain every live session's outgoing queue
}
\keyword{internal}
