% tinyrox says don't edit this manually, but it can't stop you!
\name{emit_event}
\alias{emit_event}
\title{DOM event for an emit intent}
\usage{
emit_event(emit)
}
\arguments{
\item{emit}{character "live" or "settle"}
}
\value{
character DOM event name
}
\description{
DOM event for an emit intent
}
\keyword{internal}
