% tinyrox says don't edit this manually, but it can't stop you!
\name{error_msg}
\alias{error_msg}
\title{Create an error message}
\usage{
error_msg(id, message)
}
\arguments{
\item{id}{character output ID, or NULL for session-level errors}

\item{message}{character error text}
}
\value{
character JSON string
}
\description{
Create an error message
}
\keyword{internal}
