% tinyrox says don't edit this manually, but it can't stop you!
\name{feed_declared_keep}
\alias{feed_declared_keep}
\title{The `keep` a static tree declares for this feed id, or the default}
\usage{
feed_declared_keep(id)
}
\description{
The `keep` a static tree declares for this feed id, or the default
}
\keyword{internal}
