% tinyrox says don't edit this manually, but it can't stop you!
\docType{data}
\name{FEED_KEEP_DEFAULT}
\alias{FEED_KEEP_DEFAULT}
\title{Default feed window, shared by the schema and the server-side log}
\format{
An object of class \code{integer} of length 1.
}
\usage{
FEED_KEEP_DEFAULT
}
\description{
Default feed window, shared by the schema and the server-side log
}
\keyword{internal}
