% tinyrox says don't edit this manually, but it can't stop you!
\name{find_chrome}
\alias{find_chrome}
\title{Find a browser to measure with}
\usage{
find_chrome(chrome = NULL)
}
\arguments{
\item{chrome}{character explicit binary, or NULL to search}
}
\value{
character path, or NULL when there is none
}
\description{
Find a browser to measure with
}
\keyword{internal}
