% tinyrox says don't edit this manually, but it can't stop you!
\name{fixture_json_path}
\alias{fixture_json_path}
\title{Path to the checked-in fixture JSON}
\usage{
fixture_json_path()
}
\value{
character path, or "" when the package is not installed
}
\description{
Path to the checked-in fixture JSON
}
\keyword{internal}
