% tinyrox says don't edit this manually, but it can't stop you!
\name{flitr_input}
\alias{flitr_input}
\title{Lower an input component to a flitR widget}
\usage{
flitr_input(x, session, unsupported)
}
\arguments{
\item{x}{a glinty_component}

\item{session}{the native session}

\item{unsupported}{collector env}
}
\value{
a flitR widget, or NULL
}
\description{
Lower an input component to a flitR widget
}
\keyword{internal}
