% tinyrox says don't edit this manually, but it can't stop you!
\name{flitr_output}
\alias{flitr_output}
\title{Lower an output component to a flitR item}
\usage{
flitr_output(x, values, unsupported)
}
\arguments{
\item{x}{a glinty_component}

\item{values}{env of output id -> latest value}

\item{unsupported}{collector env}
}
\value{
a flitR item, or NULL
}
\description{
Lower an output component to a flitR item
}
\keyword{internal}
