% tinyrox says don't edit this manually, but it can't stop you!
\name{head_html}
\alias{head_html}
\title{Render the head additions for a page}
\usage{
head_html(head)
}
\arguments{
\item{head}{a page head spec, or NULL}
}
\value{
character HTML (empty string when there is nothing to add)
}
\description{
Render the head additions for a page
}
\keyword{internal}
