% tinyrox says don't edit this manually, but it can't stop you!
\name{html_bind}
\alias{html_bind}
\title{Attributes that wire a control to the protocol}
\usage{
html_bind(x)
}
\arguments{
\item{x}{the component}
}
\value{
named list
}
\description{
Attributes that wire a control to the protocol
}
\keyword{internal}
