% tinyrox says don't edit this manually, but it can't stop you!
\name{html_field_group}
\alias{html_field_group}
\title{Wrap a labelled control}
\usage{
html_field_group(x, control)
}
\arguments{
\item{x}{the component}

\item{control}{character HTML of the control itself}
}
\value{
character HTML
}
\description{
Wrap a labelled control
}
\keyword{internal}
