% tinyrox says don't edit this manually, but it can't stop you!
\name{html_is_sized}
\alias{html_is_sized}
\title{Does this component carry sizing?}
\usage{
html_is_sized(x)
}
\arguments{
\item{x}{the component}
}
\value{
logical
}
\description{
Does this component carry sizing?
}
\keyword{internal}
