% tinyrox says don't edit this manually, but it can't stop you!
\name{html_outer_classes}
\alias{html_outer_classes}
\title{The classes on a rendered component's outermost element}
\usage{
html_outer_classes(html)
}
\arguments{
\item{html}{character one component's HTML}
}
\value{
character class names, empty when there are none
}
\description{
The classes on a rendered component's outermost element
}
\keyword{internal}
