% tinyrox says don't edit this manually, but it can't stop you!
\name{image_output}
\alias{image_output}
\title{Create an image output slot}
\usage{
image_output(id, alt = "")
}
\arguments{
\item{id}{character output ID}

\item{alt}{character alternative text}
}
\value{
A UI component
}
\description{
Create an image output slot
}
\examples{
image_output("preview")
}
