% tinyrox says don't edit this manually, but it can't stop you!
\name{is_input_component}
\alias{is_input_component}
\title{Is this component an input or event emitter?}
\usage{
is_input_component(name)
}
\arguments{
\item{name}{character component name}
}
\value{
logical
}
\description{
Is this component an input or event emitter?
}
\keyword{internal}
