% tinyrox says don't edit this manually, but it can't stop you!
\name{job_read_progress}
\alias{job_read_progress}
\title{Read what a job last reported}
\usage{
job_read_progress(job)
}
\arguments{
\item{job}{a glinty_job}
}
\value{
list(value, message, detail), or NULL
}
\description{
Read what a job last reported
}
\keyword{internal}
