% tinyrox says don't edit this manually, but it can't stop you!
\name{job_tail}
\alias{job_tail}
\title{The last few non-blank lines of a file}
\usage{
job_tail(path, n = 20L)
}
\arguments{
\item{path}{character file path}

\item{n}{integer how many lines}
}
\value{
character, "" when there is nothing to read
}
\description{
The last few non-blank lines of a file
}
\keyword{internal}
