% tinyrox says don't edit this manually, but it can't stop you!
\name{jwt_part_json}
\alias{jwt_part_json}
\title{Decode one JWT part as JSON}
\usage{
jwt_part_json(part)
}
\arguments{
\item{part}{character base64url segment}
}
\value{
named list, or NULL on malformed input
}
\description{
Decode one JWT part as JSON
}
\keyword{internal}
