% tinyrox says don't edit this manually, but it can't stop you!
\name{lane_running}
\alias{lane_running}
\title{How many jobs are running in a lane}
\usage{
lane_running(lane)
}
\arguments{
\item{lane}{character lane name}
}
\value{
integer
}
\description{
How many jobs are running in a lane
}
\keyword{internal}
