% tinyrox says don't edit this manually, but it can't stop you!
\name{native_text}
\alias{native_text}
\title{An output value as displayable text}
\usage{
native_text(val)
}
\arguments{
\item{val}{the value, possibly NULL}
}
\value{
character
}
\description{
An output value as displayable text
}
\keyword{internal}
