% tinyrox says don't edit this manually, but it can't stop you!
\name{native_value}
\alias{native_value}
\title{Current client-ready value of an output}
\usage{
native_value(values, id)
}
\arguments{
\item{values}{output value env}

\item{id}{character output id}
}
\value{
character value ("" before the first update)
}
\description{
Current client-ready value of an output
}
\keyword{internal}
