% tinyrox says don't edit this manually, but it can't stop you!
\name{new_condition}
\alias{new_condition}
\title{Tag a condition list}
\usage{
new_condition(x)
}
\arguments{
\item{x}{a plain list describing one condition node}
}
\value{
the list with class glinty_condition
}
\description{
Tag a condition list
}
\keyword{internal}
