% tinyrox says don't edit this manually, but it can't stop you!
\name{p}
\alias{p}
\title{Create a paragraph}
\usage{
p(..., class = NULL)
}
\arguments{
\item{...}{child elements or text}

\item{class}{character CSS class(es)}
}
\value{
A UI element
}
\description{
Create a paragraph
}
\examples{
p("Some explanatory text.")
}
