% tinyrox says don't edit this manually, but it can't stop you!
\name{parse_query}
\alias{parse_query}
\title{Parse a URL query string}
\usage{
parse_query(q)
}
\arguments{
\item{q}{character query string (no leading ?)}
}
\value{
named character vector of decoded values
}
\description{
Parse a URL query string
}
\keyword{internal}
