% tinyrox says don't edit this manually, but it can't stop you!
\name{print.glinty_tag}
\alias{print.glinty_tag}
\title{Print a tag as HTML}
\usage{
\method{print}{glinty_tag}(x, ...)
}
\arguments{
\item{x}{a glinty_tag}

\item{...}{ignored}
}
\value{
x, invisibly
}
\description{
Print a tag as HTML
}
\examples{
print(h1("Hello"))
}
