% tinyrox says don't edit this manually, but it can't stop you!
\name{progress_msg}
\alias{progress_msg}
\title{Create a progress message}
\usage{
progress_msg(action, handle)
}
\arguments{
\item{action}{character "show", "update" or "hide"}

\item{handle}{a progress handle}
}
\value{
character JSON string
}
\description{
Create a progress message
}
\keyword{internal}
