% tinyrox says don't edit this manually, but it can't stop you!
\name{real_session}
\alias{real_session}
\title{The real session behind a scoped one}
\usage{
real_session(session)
}
\arguments{
\item{session}{a session or a scoped session}
}
\value{
the glinty_session environment
}
\description{
The real session behind a scoped one
}
\keyword{internal}
