% tinyrox says don't edit this manually, but it can't stop you!
\name{schedule_flush}
\alias{schedule_flush}
\title{Schedule a flush}
\usage{
schedule_flush()
}
\description{
In webR context, sends a message to JS requesting a flush callback.
In testing context, sets a flag for manual flushing.
}
\keyword{internal}
