% tinyrox says don't edit this manually, but it can't stop you!
\name{scheme_default_port}
\alias{scheme_default_port}
\title{The port a scheme implies}
\usage{
scheme_default_port(scheme)
}
\arguments{
\item{scheme}{character lowercase scheme}
}
\value{
integer port, or NA_integer_ for schemes with none
}
\description{
The port a scheme implies
}
\keyword{internal}
