% tinyrox says don't edit this manually, but it can't stop you!
\name{slider_default}
\alias{slider_default}
\title{The HTML default for a slider with no value}
\usage{
slider_default(min, max)
}
\arguments{
\item{min,max}{numeric bounds}
}
\value{
numeric midpoint
}
\description{
The HTML default for a slider with no value
}
\keyword{internal}
