% tinyrox says don't edit this manually, but it can't stop you!
\name{span}
\alias{span}
\title{Create a generic inline element}
\usage{
span(..., class = NULL, id = NULL)
}
\arguments{
\item{...}{child elements or text}

\item{class}{character CSS class(es)}

\item{id}{character element ID}
}
\value{
A UI element
}
\description{
Create a generic inline element
}
\examples{
span("status: ok", class = "muted")
}
