% tinyrox says don't edit this manually, but it can't stop you!
\name{tab_child}
\alias{tab_child}
\title{Find a direct child of a tabset by class}
\usage{
tab_child(tg, cls)
}
\arguments{
\item{tg}{the g-tabset glinty_tag}

\item{cls}{character class to match}
}
\value{
the matching child, or NULL
}
\description{
Find a direct child of a tabset by class
}
\keyword{internal}
