% tinyrox says don't edit this manually, but it can't stop you!
\name{table_output}
\alias{table_output}
\title{Create a table output placeholder}
\usage{
table_output(id)
}
\arguments{
\item{id}{character output ID}
}
\value{
A UI element
}
\description{
Create a table output placeholder
}
\examples{
table_output("results")
}
