% tinyrox says don't edit this manually, but it can't stop you!
\name{tag_to_html}
\alias{tag_to_html}
\title{Convert a tag tree to an HTML string}
\usage{
tag_to_html(x)
}
\arguments{
\item{x}{a glinty_tag object or character}
}
\value{
character HTML string
}
\description{
Convert a tag tree to an HTML string
}
\keyword{internal}
