% tinyrox says don't edit this manually, but it can't stop you!
\name{text_output}
\alias{text_output}
\title{Create a text output placeholder}
\usage{
text_output(id)
}
\arguments{
\item{id}{character output ID}
}
\value{
A UI element
}
\description{
Create a text output placeholder
}
\examples{
text_output("greeting")
}
