% tinyrox says don't edit this manually, but it can't stop you!
\name{theme_color_parts}
\alias{theme_color_parts}
\title{Colour tokens as CSS custom property declarations}
\usage{
theme_color_parts(colors)
}
\arguments{
\item{colors}{named list of colour tokens}
}
\value{
character vector of "--g-name:value" declarations
}
\description{
Colour tokens as CSS custom property declarations
}
\keyword{internal}
