% tinyrox says don't edit this manually, but it can't stop you!
\name{theme_defaults}
\alias{theme_defaults}
\title{glinty's default look, as explicit tokens}
\usage{
theme_defaults()
}
\value{
named list of token defaults
}
\description{
glinty's default look, as explicit tokens
}
\keyword{internal}
