% tinyrox says don't edit this manually, but it can't stop you!
\name{theme_var}
\alias{theme_var}
\title{CSS custom property name for a theme token}
\usage{
theme_var(name)
}
\arguments{
\item{name}{character token name (underscores)}
}
\value{
character CSS custom property name
}
\description{
One mapping, used by both the server-side style block and the
client's welcome handler -- the token names are the contract, the
var names are the browser lowering of it.
}
\keyword{internal}
