% tinyrox says don't edit this manually, but it can't stop you!
\name{transcript_json_path}
\alias{transcript_json_path}
\title{Path to the checked-in transcript JSON}
\usage{
transcript_json_path()
}
\value{
character path, or "" when the package is not installed
}
\description{
Path to the checked-in transcript JSON
}
\keyword{internal}
