% tinyrox says don't edit this manually, but it can't stop you!
\name{translate_input}
\alias{translate_input}
\title{Translate a leaf input element}
\usage{
translate_input(tg, session, unsupported)
}
\arguments{
\item{tg}{the input glinty_tag}

\item{session}{the native session}

\item{unsupported}{collector env}
}
\value{
a flitR widget or NULL
}
\description{
Translate a leaf input element
}
\keyword{internal}
