% tinyrox says don't edit this manually, but it can't stop you!
\name{translate_tag}
\alias{translate_tag}
\title{Translate one tag}
\usage{
translate_tag(tg, session, values, unsupported)
}
\arguments{
\item{tg}{a glinty_tag or character}

\item{session}{the native session}

\item{values}{output value env}

\item{unsupported}{collector env}
}
\value{
a flitR item, or NULL to skip
}
\description{
Translate one tag
}
\keyword{internal}
