% tinyrox says don't edit this manually, but it can't stop you!
\name{translate_tags}
\alias{translate_tags}
\title{Translate a list of child tags}
\usage{
translate_tags(children, session, values, unsupported)
}
\arguments{
\item{children}{list of glinty_tag or character}

\item{session}{the native session}

\item{values}{output value env}

\item{unsupported}{collector env}
}
\value{
list of flitR items
}
\description{
Translate a list of child tags
}
\keyword{internal}
