% tinyrox says don't edit this manually, but it can't stop you!
\name{unclass_recursive}
\alias{unclass_recursive}
\title{Strip classes recursively so jsonlite serializes tag trees cleanly}
\usage{
unclass_recursive(x)
}
\arguments{
\item{x}{a UI tree or subtree}
}
\value{
the same structure with classes removed
}
\description{
Strip classes recursively so jsonlite serializes tag trees cleanly
}
\keyword{internal}
