% tinyrox says don't edit this manually, but it can't stop you!
\name{unscoped_input}
\alias{unscoped_input}
\title{An input proxy over the same environment with no namespace}
\usage{
unscoped_input(input)
}
\arguments{
\item{input}{an input proxy, scoped or not}
}
\value{
a plain glinty_input
}
\description{
An input proxy over the same environment with no namespace
}
\keyword{internal}
