% tinyrox says don't edit this manually, but it can't stop you!
\name{update_progress}
\alias{update_progress}
\title{Apply an update to a progress handle and push it}
\usage{
update_progress(handle, value, detail, message)
}
\arguments{
\item{handle}{a progress handle}

\item{value}{numeric fraction, or NULL}

\item{detail}{character, or NULL}

\item{message}{character, or NULL}
}
\value{
invisible(NULL)
}
\description{
Apply an update to a progress handle and push it
}
\keyword{internal}
