% tinyrox says don't edit this manually, but it can't stop you!
\name{ws_is_upgrade}
\alias{ws_is_upgrade}
\title{Is this request a WebSocket upgrade?}
\usage{
ws_is_upgrade(req)
}
\arguments{
\item{req}{a parsed request from parse_http_head()}
}
\value{
logical
}
\description{
Is this request a WebSocket upgrade?
}
\keyword{internal}
