% tinyrox says don't edit this manually, but it can't stop you!
\name{ws_pong_frame}
\alias{ws_pong_frame}
\title{Encode a pong frame}
\usage{
ws_pong_frame(payload = raw(0L))
}
\arguments{
\item{payload}{raw payload echoed from the ping}
}
\value{
raw frame bytes
}
\description{
Encode a pong frame
}
\keyword{internal}
