% tinyrox says don't edit this manually, but it can't stop you!
\name{children_to_flitr}
\alias{children_to_flitr}
\title{Lower a list of children, dropping the ones that contribute nothing}
\usage{
children_to_flitr(children, session, values, unsupported)
}
\arguments{
\item{children}{list of components}

\item{session}{the native session}

\item{unsupported}{collector env}
}
\value{
list of flitR items
}
\description{
Lower a list of children, dropping the ones that contribute nothing
}
\keyword{internal}
