% tinyrox says don't edit this manually, but it can't stop you!
\name{config_msg}
\alias{config_msg}
\title{Create the session config message}
\usage{
config_msg(session_id)
}
\arguments{
\item{session_id}{character session ID}
}
\value{
character JSON string
}
\description{
The first and only greeting a client receives after the WebSocket
opens.
}
\keyword{internal}
