% tinyrox says don't edit this manually, but it can't stop you!
\name{handle_download}
\alias{handle_download}
\title{Serve a registered download}
\usage{
handle_download(req)
}
\arguments{
\item{req}{a parsed request with a query carrying session and id}
}
\value{
raw HTTP response
}
\description{
Serve a registered download
}
\keyword{internal}
