% tinyrox says don't edit this manually, but it can't stop you!
\name{lane_count}
\alias{lane_count}
\title{One lane setting, as a whole number}
\usage{
lane_count(x, lane, field, min)
}
\arguments{
\item{x}{the supplied value}

\item{lane}{character lane name, for the message}

\item{field}{character field name, for the message}

\item{min}{integer smallest value that means anything}
}
\value{
integer
}
\description{
One lane setting, as a whole number
}
\keyword{internal}
