% tinyrox says don't edit this manually, but it can't stop you!
\name{new_renderer}
\alias{new_renderer}
\title{Construct a renderer}
\usage{
new_renderer(fn, property)
}
\arguments{
\item{fn}{zero-arg function producing the client-ready value}

\item{property}{character DOM property the value patches}
}
\value{
a glinty_renderer
}
\description{
Construct a renderer
}
\keyword{internal}
