% tinyrox says don't edit this manually, but it can't stop you!
\name{text_input}
\alias{text_input}
\title{Create a text input}
\usage{
text_input(id, label = "", value = "", placeholder = "")
}
\arguments{
\item{id}{character input ID}

\item{label}{character label text}

\item{value}{character initial value}

\item{placeholder}{character placeholder text}
}
\value{
A UI element
}
\description{
Create a text input
}
\examples{
text_input("name", "Name:")
}
