% tinyrox says don't edit this manually, but it can't stop you!
\name{text_output}
\alias{text_output}
\title{Create a text output slot}
\usage{
text_output(id, variant = "normal")
}
\arguments{
\item{id}{character output ID}

\item{variant}{character "normal", "muted" or "strong"}
}
\value{
A UI component
}
\description{
Create a text output slot
}
\examples{
text_output("greeting")
}
